fix: resolve mypy type errors (#4)
* chore: rename project to FlowHawk, add CI workflow

- Rename package from quant-research to flowhawk in pyproject.toml
- Add GitHub Actions CI with ruff, mypy, pytest
- CI triggers on push/PR to main and develop branches

* fix: resolve ruff lint errors

- Fix F541: remove f-string without placeholders in fmp.py
- Fix F821: add missing Optional import in yfinance_ds.py, stock_technical.py
- Fix F401: remove unused imports (auto-fixed by ruff)
- Add __init__.py to dashboard/, scripts/, tests/ for git tracking

* style: apply ruff format to all source files

- Format 15 files with ruff format
- All lint checks and format checks now pass

* fix: resolve mypy type errors

- Fix valid-type error: rename date to Date in trade_signal.py
- Fix return-type error: cast numpy float to float in stock_technical.py
- Fix operator error: add type ignore for polars expr
- Fix call-overload: add isinstance check in fmp.py

---------

+++ src/screening/stock_technical.py
@@ -137,7 +137,7 @@ class StockTechnicalScreener:
         if avg_loss == 0:
             return 100.0
         rs = avg_gain / avg_loss
-        return 100.0 - (100.0 / (1.0 + rs))
+        return float(100.0 - (100.0 / (1.0 + rs)))
 
     @staticmethod
     def _compute_atr(df: pl.DataFrame, period: int = 14) -> float:
